requirements.txt
flask==3.1.2
pytesseract==0.3.13
pillow==11.3.0
transformers==4.57.3
torch==2.8.0